% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teda-b-s3-plot.R
\name{plot.tedab}
\alias{plot.tedab}
\title{Plot the tedab object}
\usage{
\method{plot}{tedab}(x, ...)
}
\arguments{
\item{x}{The teda batch (tedab) object with which to create the plot output.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Takes a tedab object and plots each metric individually
}
\details{
Takes a tedab object and creates four plots in order of:
eccentricity, typicality, normalised eccentricity, and normalised typicality.
}

