% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teda-b-s3-print.R
\name{print.tedab}
\alias{print.tedab}
\title{Print the tedab object}
\usage{
\method{print}{tedab}(x, ...)
}
\arguments{
\item{x}{The teda batch (tedab) object with which to create the printed output.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Takes a tedab object and prints out the values within
}
\details{
Takes a tedab object and prints out each vector in order of:
eccentricity, typicality, normalised eccentricity, and normalised typicality.
}

