% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teda-r-s3-print.R
\name{print.tedar}
\alias{print.tedar}
\title{Print the tedar object}
\usage{
\method{print}{tedar}(x, ...)
}
\description{
Takes a tedar object and prints out the values within (currently the same as summarize).
}

