% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teda-b-s3-summary.R
\name{summary.tedab}
\alias{summary.tedab}
\title{Summarise the tedab object}
\usage{
\method{summary}{tedab}(x, ...)
}
\description{
Takes a tedab object and prints out the following summary details:
\itemize{
 \item{the number of observations}
 \item{the number of observations that exceed the normalised eccentricity limit}
 \item{the normalised eccentricity threshold}
}
}

