% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teda.R
\docType{package}
\name{teda}
\alias{teda}
\alias{teda-package}
\title{teda: An implementation of the Typicality and Eccentricity Data Analysis
    Framework.}
\description{
The package provides functions to calculate both the batch and recursive
    typicality and eccentricity values of given observations.
}
\details{
TEDA provides a non-parametric technique to determine how eccentric/typical
an observation is with respect to the other observations generated by the
same process.  Available as either a batch function working over a whole
dataset, or as a recursive one-time-pass function that needs the current
mean and variance values to be passed as arguments.

Both batch and recursive methods return a datatype (tedab or tedar) which
provide print and summary generic function implementations.  The batch
object also provides a generic plot function.

Further work will implement more of the analytical framework built up
around TEDA, such as clustering algorithms.
}
\references{
Angelov, P., 2014. Outside the box: an alternative data analytics framework.
Journal of Automation Mobile Robotics and Intelligent Systems, 8(2), pp.29-35.
DOI: 10.14313/JAMRIS_2-2014/16

Bezerra, C.G., Costa, B.S.J., Guedes, L.A. and Angelov, P.P., 2016, May.
A new evolving clustering algorithm for online data streams.
In Evolving and Adaptive Intelligent Systems (EAIS),
2016 IEEE Conference on (pp. 162-168). IEEE.
DOI: 10.1109/EAIS.2016.7502508
}

