% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teda-r-s3-print.R
\name{print.tedar}
\alias{print.tedar}
\title{Print the tedar object}
\usage{
\method{print}{tedar}(x, ...)
}
\arguments{
\item{x}{The teda recursive (tedar) object with which to create the print output.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
#' @description Takes a tedar object and prints out the object values.
}
\details{
Takes a tedar object and prints out the values within (currently the same as summarize).
}

